% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhD.R
\name{make_chairmen}
\alias{make_chairmen}
\title{Make chairmen duos}
\usage{
make_chairmen(names)
}
\arguments{
\item{names}{character vector with names of aios from Tytgat (both floors)}
}
\value{
matrix of chairmen couples
}
\description{
For the Thursday morning meeting, make duos that will chair the meeting consisting of only aios from the Tytgat.
}
