% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhD.R
\name{PhD_life}
\alias{PhD_life}
\title{Fun function for PhDs!}
\usage{
PhD_life()
}
\value{
a meme saved into your environment
}
\description{
Have a break from the serious coding and use this function to cheer yourself up with a meme.
}
