% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tytgat_aio.R
\docType{data}
\name{tytgat_aio}
\alias{tytgat_aio}
\title{Tytgat aio names}
\format{
character vector
}
\source{
List of all the PhDs of the Tytgat. This will need to be updated once in a while.
}
\usage{
tytgat_aio
}
\description{
A (character) vector consisting of the names of all the Tytgat aios.
}
\keyword{datasets}
